+++ src/proxy/inbound_passthrough.rs
@@ -16,10 +16,11 @@ use std::net::SocketAddr;
 
 use drain::Watch;
 use tokio::net::{TcpListener, TcpStream};
+use tokio::sync::watch;
 use tracing::{error, info, trace, warn, Instrument};
 
 use crate::config::ProxyMode;
-use crate::proxy::connection_manager::ConnectionManager;
+use crate::proxy::connection_manager::{self, ConnectionManager};
 use crate::proxy::metrics::Reporter;
 use crate::proxy::outbound::OutboundConnection;
 use crate::proxy::{metrics, util, ProxyInputs};
@@ -64,11 +65,24 @@ impl InboundPassthrough {
     }
 
     pub(super) async fn run(self) {
+        // spawn a task which subscribes to watch updates and asserts rbac against this proxy's connections, closing the ones which have become denied
+        let (stop_tx, stop_rx) = watch::channel(());
+        let connection_manager = self.connection_manager.clone();
+        let state = self.pi.state.clone();
+
+        tokio::spawn(connection_manager::policy_watcher(
+            state,
+            stop_rx,
+            connection_manager,
+            "inbound_passthrough",
+        ));
         let accept = async move {
         loop {
             // Asynchronously wait for an inbound socket.
             let socket = self.listener.accept().await;
             let pi = self.pi.clone();
+
+
             let connection_manager = self.connection_manager.clone();
             match socket {
                 Ok((stream, remote)) => {
@@ -77,7 +91,7 @@ impl InboundPassthrough {
                             pi, // pi cloned above; OK to move
                             socket::to_canonical(remote),
                             stream,
-                                connection_manager,
+                            connection_manager,
                         )
                         .await
                         {
@@ -94,7 +108,6 @@ impl InboundPassthrough {
             }
         }
       }.in_current_span();
-
         // Stop accepting once we drain.
         // Note: we are *not* waiting for all connections to be closed. In the future, we may consider
         // this, but will need some timeout period, as we have no back-pressure mechanism on connections.
@@ -102,6 +115,7 @@ impl InboundPassthrough {
             res = accept => { res }
             _ = self.drain.signaled() => {
                 info!("inbound passthrough drained");
+                stop_tx.send_replace(());
             }
         }
     }
